done @ check build

+++ src/app/api/accessManagements/v1/route.ts
@@ -14,19 +14,19 @@ import {
   //    handleTokenVerify,
   IResponse,
 } from '@/app/api/utils/jwt-verify';
-// import { isUserHasAccess, IWantAccess } from '@/app/api/utils/is-user-has-access';
+import { isUserHasAccessByRole, IWantAccess } from '../../utils/is-user-has-access-by-role';
 
 // GET all AccessManagements
 export async function GET(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'access_management',
-  //   access: 'read',
-  // };
-  // const isAccess = await isUserHasAccess(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'access',
+    access: 'read',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const id = new URL(req.url).searchParams.get('id');
   const result: IResponse = id ? await getAccessManagementById(req) : await getAccessManagements(req);
@@ -38,12 +38,12 @@ export async function POST(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'access_management',
-  //   access: 'create',
-  // };
-  // const isAccess = await isUserHasAccess(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'access',
+    access: 'create',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const result = await createAccessManagement(req);
   return formatResponse(result.data, result.message, result.status);
@@ -54,12 +54,12 @@ export async function PUT(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'access_management',
-  //   access: 'update',
-  // };
-  // const isAccess = await isUserHasAccess(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'access',
+    access: 'update',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const isBulk = new URL(req.url).searchParams.get('bulk') === 'true';
   const result = isBulk ? await bulkUpdateAccessManagements(req) : await updateAccessManagement(req);
@@ -72,12 +72,12 @@ export async function DELETE(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'access_management',
-  //   access: 'delete',
-  // };
-  // const isAccess = await isUserHasAccess(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'access',
+    access: 'delete',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const isBulk = new URL(req.url).searchParams.get('bulk') === 'true';
   const result = isBulk ? await bulkDeleteAccessManagements(req) : await deleteAccessManagement(req);
