+++ src/app/api/media/v1/route.ts
@@ -1,18 +1,18 @@
 import { NextResponse } from 'next/server';
 import { getMedia, createMedia, updateMedia, deleteMedia, getMediaById } from './controller';
 import { handleRateLimit } from '@/app/api/utils/rate-limit';
-// import { isUserHasAccessByRole, IWantAccess } from '../../utils/is-user-has-access-by-role';
+import { isUserHasAccessByRole, IWantAccess } from '../../utils/is-user-has-access-by-role';
 
 export async function GET(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'media',
-  //   access: 'read',
-  // };
-  // const isAccess = await isUserHasAccessByRole(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'media',
+    access: 'read',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const id = new URL(req.url).searchParams.get('id');
   const result = id ? await getMediaById(req) : await getMedia(req);
@@ -24,12 +24,12 @@ export async function POST(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'media',
-  //   access: 'create',
-  // };
-  // const isAccess = await isUserHasAccessByRole(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'media',
+    access: 'create',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const result = await createMedia(req);
   return NextResponse.json(result.data, { status: result.status, statusText: result.message });
@@ -39,12 +39,12 @@ export async function PUT(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'media',
-  //   access: 'update',
-  // };
-  // const isAccess = await isUserHasAccessByRole(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'media',
+    access: 'update',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const result = await updateMedia(req);
   return NextResponse.json(result.data, { status: result.status, statusText: result.message });
@@ -54,12 +54,12 @@ export async function DELETE(req: Request) {
   const rateLimitResponse = handleRateLimit(req);
   if (rateLimitResponse) return rateLimitResponse;
 
-  // const wantToAccess: IWantAccess = {
-  //   db_name: 'media',
-  //   access: 'delete',
-  // };
-  // const isAccess = await isUserHasAccessByRole(wantToAccess);
-  // if (isAccess) return isAccess;
+  const wantToAccess: IWantAccess = {
+    db_name: 'media',
+    access: 'delete',
+  };
+  const isAccess = await isUserHasAccessByRole(wantToAccess);
+  if (isAccess) return isAccess;
 
   const result = await deleteMedia(req);
   return NextResponse.json(result.data, { status: result.status, statusText: result.message });
